+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Changelog for NeoFS Contract
 
 ### Removed
 - neofsid contract (#474)
+- bind/unbind methods of neofs contract (#475)
 
 ### Fixed